third_party/blink/web_tests/external/wpt/css/css-values/container-progress-computed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] container-progress(height, 0px, 1px) should be used-value-equivalent to 322
  assert_equals: container-progress(height, 0px, 1px) and 322 serialize to the same thing in used values. expected "322" but got "1"
[FAIL] container-progress(width of my-container, 0px, 1px) should be used-value-equivalent to 322
  assert_equals: container-progress(width of my-container, 0px, 1px) and 322 serialize to the same thing in used values. expected "322" but got "1"
[FAIL] container-progress(height, container-progress(height, 0px, 1px) * 1px, container-progress(height, 0px, 1px) * 1px) should be used-value-equivalent to 0
  assert_equals: container-progress(height, container-progress(height, 0px, 1px) * 1px, container-progress(height, 0px, 1px) * 1px) and 0 serialize to the same thing in used values. expected "0" but got "1"
[FAIL] container-progress(height, container-progress(width of my-container, 0px, 1px) * 1px, container-progress(height of my-container-2, 0px, 1px) * 1px) should be used-value-equivalent to 0
  assert_equals: container-progress(height, container-progress(width of my-container, 0px, 1px) * 1px, container-progress(height of my-container-2, 0px, 1px) * 1px) and 0 serialize to the same thing in used values. expected "0" but got "1"
[FAIL] calc(container-progress(width, 0px, 50px) * 10px + 100px) should be used-value-equivalent to 145.6px
  assert_equals: calc(container-progress(width, 0px, 50px) * 10px + 100px) and 145.6px serialize to the same thing in used values. expected "145.6px" but got "110px"
[FAIL] calc(container-progress(height, 10px, sign(50px - 500em) * 10px)) should be used-value-equivalent to -15.6
  assert_equals: calc(container-progress(height, 10px, sign(50px - 500em) * 10px)) and -15.6 serialize to the same thing in used values. expected "-15.6" but got "0"
[FAIL] calc(container-progress(width of my-container, 0px, 50px) * 10px + 100px) should be used-value-equivalent to 164.4px
  assert_equals: calc(container-progress(width of my-container, 0px, 50px) * 10px + 100px) and 164.4px serialize to the same thing in used values. expected "164.4px" but got "110px"
[FAIL] calc(container-progress(height of my-container, 10px, sign(50px - 500em) * 10px)) should be used-value-equivalent to -10.9
  assert_equals: calc(container-progress(height of my-container, 10px, sign(50px - 500em) * 10px)) and -10.9 serialize to the same thing in used values. expected "-10.9" but got "0"
[FAIL] container-progress() width fallback for non-existing container name
  assert_equals: container-progress(width of non-existing-container, 0px, 1px) and 800 serialize to the same thing in used values. expected "800" but got "1"
[FAIL] container-progress() height fallback for non-existing container names
  assert_equals: container-progress(height of non-existing-container, 0px, 1px) and 600 serialize to the same thing in used values. expected "600" but got "1"
[FAIL] container-progress() width fallback for out of scope container
  assert_equals: container-progress(width of out-of-scope-container, 0px, 1px) and 800 serialize to the same thing in used values. expected "800" but got "1"
[FAIL] container-progress() height fallback for out of scope container
  assert_equals: container-progress(height of out-of-scope-container, 0px, 1px) and 600 serialize to the same thing in used values. expected "600" but got "1"
[FAIL] calc(container-progress(width, 0px, 1px) * 1px) should be used-value-equivalent to 228px
  assert_equals: calc(container-progress(width, 0px, 1px) * 1px) and 228px serialize to the same thing in used values. expected "228px" but got "1px"
[FAIL] calc(container-progress(height of my-container, 0px, 1px) * 1s) should be used-value-equivalent to 228s
  assert_equals: calc(container-progress(height of my-container, 0px, 1px) * 1s) and 228s serialize to the same thing in used values. expected "228s" but got "1s"
[FAIL] calc(container-progress(width of my-container-2, 0px, 1px) * 1deg) should be used-value-equivalent to 5051deg
  assert_approx_equals: calc(container-progress(width of my-container-2, 0px, 1px) * 1deg) and 5051deg 0.001 serialize to the same thing in used values. expected 5051 +/- 0.001 but got 1
Harness: the test ran to completion.
